More longs fixed (should have no effect).
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7005 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/DarkModGlobals.h
@@ -216,7 +216,7 @@ public:
 
 	LC_LogClass		m_LogClass;
 	LT_LogType		m_LogType;
-	long			m_Frame;
+	int			m_Frame;
 	const char		*m_Filename;
 	char			m_DriveLetter;		// Remember the last driveletter
 	int				m_Linenumber;
