+++ game/gamesys/Class.cpp
@@ -280,7 +280,7 @@ idClass::FindUninitializedMemory
 */
 void idClass::FindUninitializedMemory( void ) {
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-	unsigned long *ptr = ( ( unsigned long * )this ) - 1;
+	unsigned int *ptr = ( ( unsigned int * )this ) - 1;
 	int size = *ptr;
 	assert( ( size & 3 ) == 0 );
 	size >>= 2;
@@ -443,13 +443,13 @@ void * idClass::operator new( size_t s ) {
 	int *p;
 
 	s += sizeof( int );
-    p = (int *)Mem_Alloc(static_cast<int>(s));
-    *p = static_cast<int>(s);
-    memused += static_cast<int>(s);
+	p = (int *)Mem_Alloc(static_cast<int>(s));
+	*p = static_cast<int>(s);
+	memused += static_cast<int>(s);
 	numobjects++;
 
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-    unsigned long *ptr = (unsigned long *)p;
+	unsigned int *ptr = (unsigned int *)p;
 	int size = s;
 	assert( ( size & 3 ) == 0 );
 	size >>= 3;
@@ -465,13 +465,13 @@ void * idClass::operator new( size_t s, int, int, char *, int ) {
 	int *p;
 
 	s += sizeof( int );
-    p = (int *)Mem_Alloc(static_cast<int>(s));
-    *p = static_cast<int>(s);
-    memused += static_cast<int>(s);
+	p = (int *)Mem_Alloc(static_cast<int>(s));
+	*p = static_cast<int>(s);
+	memused += static_cast<int>(s);
 	numobjects++;
 
 #ifdef ID_DEBUG_UNINITIALIZED_MEMORY
-    unsigned long *ptr = (unsigned long *)p;
+	unsigned int *ptr = (unsigned int *)p;
 	int size = s;
 	assert( ( size & 3 ) == 0 );
 	size >>= 3;
